Initial version of nvgrid element
It is possible to build image of ubuntu using this elements and it
has nvidia drivers with the changes from the manual.

There is a two variable, which must be defined in order to proceed
with the build: DIB_DRIVER_URL and DIB_CLIENT_TOKEN The first one
should include url to download deb package with the driver. The
second is a filename of the token.