Add hook around systemd-resolved installation
Once systemd-resolved is installed, it replaces /etc/resolv.conf with
a symlink to /run, where resolved is supposed to place resulting
configuration.

However, it is not possible to startup a service inside chroot
environment. So, we workaround that by preserving resolv.conf
content in destination file for resolved before it is installed.